python-processor/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.6
minio==7.1.14
python-dateutil==2.8.2